% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHL_Schedule.R
\name{get_Schedule}
\alias{get_Schedule}
\title{Get the Current OHL Schedule}
\usage{
get_Schedule()
}
\value{
A data frame containing the current OHL regular season schedule.
}
\description{
Retrieves the current OHL regular season schedule.
}
\examples{
\dontrun{
schedule <- get_Schedule()
head(schedule)
}

}
