% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHL_Schedule.R
\name{get_Schedule}
\alias{get_Schedule}
\title{Get OHL schedule}
\usage{
get_Schedule(Schedule)
}
\value{
A data frame with the OHL regular season schedule.
}
\description{
Retrieves the OHL regular season schedule for the current season.
}
\examples{
sched <- get_Schedule()
head(sched)
}
