% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OHL_Teams.R
\name{get_Teams}
\alias{get_Teams}
\title{Get OHL teams}
\usage{
get_Teams(Teams)
}
\value{
A data frame with OHL team details.
}
\description{
Retrieves a data frame with team information for the 2024–2025 or 2025–2026 season.
}
\examples{
teams <- get_Teams()
head(teams)
}
