\name{Hello}
\alias{Hello}
\title{OHLpkg!}
\usage{
get_Stats()
get_EVStats()
}
\description{
get_Stats() creates a dataframe with the 2023-2024 Player Statistics.
get_EVStats() creates a dataframe with the 2023-2024 EV Strength Player Stats.
Both of these dataframes have filtered out players who haven't played at least 10 games.
}
\examples{
get_Stats()
get_EVStats()
}
